backend/requirements/base.txt
# HRMS Backend Requirements

# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14
pytz>=2024.1
django-cors-headers>=4.3
django-filter>=23.5
djangorestframework-simplejwt>=5.3

# WebSockets / Async
daphne>=4.0
channels>=4.0
channels_redis>=4.0

# Multi-tenancy (Organization-based)
# django-tenants removed

# Encryption
cryptography>=41.0
django-encrypted-model-fields>=0.6

# Database
psycopg2-binary>=2.9

# Async Tasks
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5

# API Documentation
drf-spectacular>=0.27

# Testing
factory-boy>=3.3
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1

# Logging
python-json-logger>=2.0.7

# Security
cryptography>=42.0.0
django-axes>=6.3
django-otp>=1.2
qrcode>=7.4

# GraphQL
strawberry-graphql-django>=0.38

# Search
django-elasticsearch-dsl>=8.0

# Storage
boto3>=1.34
django-storages>=1.14

# AI/ML
openai>=1.10
langchain>=0.1
sentence-transformers>=2.3
spacy>=3.7

# Utils
python-dateutil>=2.8
python-decouple>=3.8
Pillow>=10.2
python-magic>=0.4
pandas>=2.2
openpyxl>=3.1
reportlab>=4.1
WeasyPrint>=61.0

# Monitoring
sentry-sdk>=1.40
prometheus-client>=0.19

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.1
factory-boy>=3.3
faker>=22.5

# Code Quality
black>=24.1
isort>=5.13
flake8>=7.0
mypy>=1.8
pyotp>=2.8
djangorestframework-stubs
